Automatic task name detection
Detect task names inside gulp tasks if the task name is being quoted in the console. This is most frequently used with things such as `$.size()` where the task name is used for informational and debugging purposes where repeated hard coding of the task name can lead to errors caused by failure to update all references to the correct task name. Ref #19